refactor(web): move workspace analytics into views/admin and model chart state
