+++ web/src/views/admin/WorkspaceAnalyticsPage/index.tsx
@@ -1,0 +1,51 @@
+"use client";
+
+import { AdminDateRangeSelector } from "@/components/dateRangeSelectors/AdminDateRangeSelector";
+import { useTimeRange } from "@/app/ee/admin/performance/lib";
+import {
+  FeedbackChart,
+  SlackChannelChart,
+  UsageChart,
+} from "@/views/admin/WorkspaceAnalyticsPage/AnalyticsCharts";
+import { PersonaMessagesChart } from "@/views/admin/WorkspaceAnalyticsPage/PersonaMessagesChart";
+import UsageReports from "@/views/admin/WorkspaceAnalyticsPage/UsageReports";
+import { ADMIN_ROUTES } from "@/lib/admin-routes";
+import { Divider } from "@opal/components";
+import { Section, SettingsLayouts } from "@opal/layouts";
+
+const route = ADMIN_ROUTES.WORKSPACE_ANALYTICS;
+
+export default function WorkspaceAnalyticsPage() {
+  const [timeRange, setTimeRange] = useTimeRange();
+
+  return (
+    <SettingsLayouts.Root width="lg">
+      <SettingsLayouts.Header
+        icon={route.icon}
+        title={route.title}
+        description="Understand how your workspace uses Onyx across queries, feedback, and agents."
+        divider
+      />
+      <SettingsLayouts.Body>
+        <Section flexDirection="row" justifyContent="end" height="fit">
+          <AdminDateRangeSelector
+            value={timeRange}
+            onValueChange={(range) =>
+              setTimeRange((previous) =>
+                range
+                  ? { ...range, selectValue: previous.selectValue }
+                  : previous
+              )
+            }
+          />
+        </Section>
+        <UsageChart timeRange={timeRange} />
+        <FeedbackChart timeRange={timeRange} />
+        <SlackChannelChart timeRange={timeRange} />
+        <PersonaMessagesChart timeRange={timeRange} />
+        <Divider />
+        <UsageReports />
+      </SettingsLayouts.Body>
+    </SettingsLayouts.Root>
+  );
+}
